refactor loginUser on refresh
