+++ frontend/src/pages/Home.vue
@@ -20,19 +20,6 @@ export default {
   mounted() {
     const fenixCode = this.$route.query.code;
     if (fenixCode) this.authBackend(fenixCode);
-
-    const jwt = localStorage.getItem('token');
-    if (jwt) {
-      const res = parseJwt(jwt);
-      const user = {
-        id: res.id,
-        istId: res.istId,
-        name: res.name,
-        active: res.active,
-        admin: res.admin,
-      };
-      this.loginUser(user);
-    }
   },
   methods: {
     login() {
@@ -67,20 +54,4 @@ export default {
     ...mapActions('user', ['loginUser', 'logoutUser']),
   },
 };
-
-//gets the user data from the middle section of the jwt if its already present in local storage
-function parseJwt(token) {
-  var base64Url = token.split('.')[1];
-  var base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
-  var jsonPayload = decodeURIComponent(
-    atob(base64)
-      .split('')
-      .map(function (c) {
-        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
-      })
-      .join('')
-  );
-
-  return JSON.parse(jsonPayload);
-}
 </script>
